fix: improve error handling non-existing output dir
This commit addresses the following two issues:
* when the output directory given to the base extracter does not exist,
the error message displayed complains that the file being created does
not exist, while this is not the actual issue.
* When the output directory is not empty, the files contained in it will
be overwritten without a warning.

I addressed this by checking first if the output directory exists and is
empty, if not, the appropriate warning or error message will be
raised/displayed.