Add example directory
Add an example directory that provides a script to download the
Gulordava model and use it in the scripts we provided.

This is a very first version, the folder isn't quite finished yet, both
in terms of README as in terms of example scripts. Currentlty, there is
only the script `extract.py`, and its corresponding config file. The
model is downloaded, but the data to run it on is just provided (since
we don't have it hosted somewher online).

More expanation and scripts should be added, preferably also some
scripts that allow to transform simple tab separated files to labeled
files.

I was also thinking we may put this in travis, so that it checks if the
scripts work.